refactor: decompose postgres.py into per-repo modules
Extract 5 repository classes from postgres.py (798 lines):
- postgres_player.py: PostgresPlayerRepository (80 lines)
- postgres_session.py: PostgresSessionRepository (77 lines)
- postgres_game.py: PostgresGameRepository (207 lines)
- postgres_turn.py: PostgresTurnRepository (227 lines)
- postgres_world_event.py: PostgresWorldEventRepository (96 lines)

postgres.py → 176 lines (convenience stubs + backward-compat re-exports).
All classes independently importable. No cross-class dependencies.
